feat GameConsole 객체 추가 (게임 실행)
refactor Participants 클래스에서 게임 제어 책임 제거 -> GameConsole로 위임
