+++ src/main/java/ladder/README.md
@@ -116,6 +116,8 @@
     - Map<Person,Result>를 가지는 일급 컬렉션
     - 전체 게임 결과를 가진다 
     
+11. __GameConsole 게임 실행__
+    - 사다리 게임을 실행하는 역할을 맡는다
 
 자바 8의 스트림과 람다를 적용해 프로그래밍한다.  
 규칙 6: 모든 엔티티를 작게 유지한다.
