+++ src/main/java/ladder/domain/GameConsole.java
@@ -1,0 +1,32 @@
+package ladder.domain;
+
+import java.util.LinkedHashMap;
+import java.util.Map;
+
+public class GameConsole {
+    private Participants participants;
+    private Ladder ladder;
+
+    private GameConsole(Participants participants, Ladder ladder) {
+        this.participants = participants;
+        this.ladder = ladder;
+    }
+
+    public static GameConsole valueOf(Participants participants, Ladder ladder) {
+        return new GameConsole(participants, ladder);
+    }
+
+
+    public GameResult runLadder(Results results) {
+        Map<Person, Result> gameResult = new LinkedHashMap<>();
+        int personIndex = 0;
+        for (Person person : participants.getParticipants()) {
+            int finalPoint = ladder.run(personIndex);
+            Result result = results.find(finalPoint);
+            gameResult.put(person, result);
+            personIndex++;
+        }
+        return GameResult.valueOf(gameResult);
+    }
+
+}
